<==== Mercurial data ====>
Branch: release_2.3

+++ common/src/core/resource/camera_resource.cpp
@@ -29,14 +29,14 @@ QnPhysicalCameraResource::QnPhysicalCameraResource():
 
 int QnPhysicalCameraResource::suggestBitrateKbps(Qn::StreamQuality quality, QSize resolution, int fps) const
 {
-    float lowEnd = 0.1;
-    float hiEnd = 1.0;
+    float lowEnd = 0.1f;
+    float hiEnd = 1.0f;
 
     float qualityFactor = lowEnd + (hiEnd - lowEnd) * (quality - Qn::QualityLowest) / (Qn::QualityHighest - Qn::QualityLowest);
 
-    float resolutionFactor = 0.009 * pow(resolution.width() * resolution.height(), (float)0.7);
+    float resolutionFactor = 0.009f * pow(resolution.width() * resolution.height(), 0.7f);
 
-    float frameRateFactor = fps/1.0;
+    float frameRateFactor = fps/1.0f;
 
     float result = qualityFactor*frameRateFactor * resolutionFactor;
